+++ rustfs/src/app/mod.rs
@@ -31,6 +31,8 @@ mod capacity_dirty_scope_test;
 #[cfg(test)]
 mod delete_objects_stat_gating_test;
 #[cfg(test)]
+mod gating_test_env;
+#[cfg(test)]
 mod lifecycle_transition_api_test;
 #[cfg(test)]
 mod put_prelookup_gating_test;
